Moving request-mocking to its own file.
